requirements.txt
apt install build-essential libssl-dev libffi-dev python3-dev
apt install nmap
sudo apt-get install chromium-browser

aiohttp==3.6.2
aiomultiprocess==0.7.0
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
Brotli==1.0.7
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
cloudscraper==1.2.33
colorama==0.4.3
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.0.1
exrex==0.10.5
fire==0.3.1
idna==2.9
jdcal==1.4.1
Js2Py==0.70
loguru==0.4.1
multidict==4.7.5
openpyxl==2.4.11
psutil==5.7.0
pyjsparser==2.7.1
PySocks==1.7.1
pytz==2019.3
records==0.5.3
requests==2.23.0
requests-file==1.4.3
requests-toolbelt==0.9.1
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
tablib==1.1.0
tenacity==6.1.0
termcolor==1.1.0
tldextract==2.2.2
tqdm==4.45.0
tzlocal==2.0.0
urllib3==1.25.9
win32-setctime==1.0.1
yarl==1.4.2

flask
flask-mail
Flask-SQLAlchemy
SQLAlchemy
Flask-RESTful
ipip-ipdb
psycopg2-binary
shodan
python-nmap
pluginbase